# testdata/project/golden/minimal/impl/echoserver/server.gen.go
// Code generated by oapigen, DO NOT EDIT.

package server

import (
	"bytes"
	"encoding/json"
	"github.com/labstack/echo/v4"
	"io"
	"net/http"
	apiclient "nschugorev/oapigenerator/testdata/project/golden/minimal/interfaces/client"
	apiserver "nschugorev/oapigenerator/testdata/project/golden/minimal/interfaces/server"
)

type ServerHTTP struct {
	impl apiserver.Server
}

func NewServerHTTP(impl apiserver.Server) *ServerHTTP {
	return &ServerHTTP{impl: impl}
}

// bindBody читает тело запроса, сбрасывает его для c.Bind() и
// десериализует в dst. c.Bind() повторно читает тело, но поля с
// json:"-" игнорируются — поэтому Body декодируется отдельно.
func bindBody(c echo.Context, dst any) error {
	body, err := io.ReadAll(c.Request().Body)
	if err != nil {
		return err
	}
	c.Request().Body = io.NopCloser(bytes.NewReader(body))
	if len(body) == 0 {
		return nil
	}
	if err := json.Unmarshal(body, dst); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	return nil
}

func (s *ServerHTTP) Register(e *echo.Echo) {
	e.GET("/items", s.listItems)
	e.POST("/items", s.createItem)
	e.GET("/items/:id", s.getItem)
	e.PUT("/items/:id", s.updateItem)
	e.DELETE("/items/:id", s.deleteItem)
}

func (s *ServerHTTP) listItems(c echo.Context) error {
	req := &apiclient.ListItemsRequest{}
	if err := c.Bind(req); err != nil {
		return err
	}
	resp, err := s.impl.ListItems(c.Request().Context(), req)
	if err != nil {
		return err
	}
	if resp.Response200 != nil {
		return c.JSON(200, resp.Response200)
	}
	if resp.Response400 {
		return c.NoContent(400)
	}
	return c.NoContent(resp.Code)
}

func (s *ServerHTTP) createItem(c echo.Context) error {
	req := &apiclient.CreateItemRequest{}
	if err := bindBody(c, &req.Body); err != nil {
		return err
	}
	if err := c.Bind(req); err != nil {
		return err
	}
	resp, err := s.impl.CreateItem(c.Request().Context(), req)
	if err != nil {
		return err
	}
	if resp.Response201 != nil {
		return c.JSON(201, resp.Response201)
	}
	if resp.Response400 {
		return c.NoContent(400)
	}
	return c.NoContent(resp.Code)
}

func (s *ServerHTTP) getItem(c echo.Context) error {
	req := &apiclient.GetItemRequest{}
	if err := c.Bind(req); err != nil {
		return err
	}
	resp, err := s.impl.GetItem(c.Request().Context(), req)
	if err != nil {
		return err
	}
	if resp.Response200 != nil {
		return c.JSON(200, resp.Response200)
	}
	if resp.Response404 {
		return c.NoContent(404)
	}
	return c.NoContent(resp.Code)
}

func (s *ServerHTTP) updateItem(c echo.Context) error {
	req := &apiclient.UpdateItemRequest{}
	if err := bindBody(c, &req.Body); err != nil {
		return err
	}
	if err := c.Bind(req); err != nil {
		return err
	}
	resp, err := s.impl.UpdateItem(c.Request().Context(), req)
	if err != nil {
		return err
	}
	if resp.Response200 != nil {
		return c.JSON(200, resp.Response200)
	}
	if resp.Response400 {
		return c.NoContent(400)
	}
	return c.NoContent(resp.Code)
}

func (s *ServerHTTP) deleteItem(c echo.Context) error {
	req := &apiclient.DeleteItemRequest{}
	if err := c.Bind(req); err != nil {
		return err
	}
	resp, err := s.impl.DeleteItem(c.Request().Context(), req)
	if err != nil {
		return err
	}
	if resp.Response204 {
		return c.NoContent(204)
	}
	return c.NoContent(resp.Code)
}