# testdata/project/golden/minimal/model/created_event.gen.go
// Code generated by oapigen, DO NOT EDIT.

package model

import (
	"fmt"
	validator "nschugorev/oapigenerator/pkg/validator"
)

type CreatedEvent struct {
	Type   string `json:"type" yaml:"type"`
	ItemID string `json:"itemId" yaml:"itemId"`
}

func (x CreatedEvent) ValidateOwn(reg *validator.Registry) error {
	if len(x.ItemID) < 1 {
		return fmt.Errorf("field ItemID: must be >= 1")
	}
	return nil
}