# testdata/project/minimal/src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: minimal
  version: '1.0.0'
paths:
  /items:
    get:
      operationId: listItems
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
        - {name: tag, in: query, schema: {type: string}}
      responses:
        '200':
          description: list of items
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Items'}
        '400': {description: bad request}
    post:
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ItemCreate'}
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Item'}
        '400': {description: bad request}
  /items/{id}:
    get:
      operationId: getItem
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: item
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Item'}
        '404': {description: not found}
    put:
      operationId: updateItem
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ItemCreate'}
      responses:
        '200':
          description: updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Item'}
        '400': {description: bad request}
    delete:
      operationId: deleteItem
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '204': {description: deleted}
components:
  schemas:
    Item:
      type: object
      required: [id, name, kind]
      x-validations: ["app.ItemConsistency"]
      properties:
        id:
          type: string
          x-validations: ["Size >=1"]
        name:
          type: string
          x-validations: ["Size >=1"]
        kind: {$ref: '#/components/schemas/Kind'}
        tag:
          type: string
          nullable: true
          x-validations: ["Size <=50"]
        secret:
          type: string
          x-sensitive: true
        metadata: {type: object, additionalProperties: {type: string}}
    ItemCreate:
      type: object
      required: [name, kind]
      x-validations: ["app.ItemCreateConsistency"]
      properties:
        name:
          type: string
          x-validations: ["Size >=1", "app.NonEmptyName"]
        kind: {$ref: '#/components/schemas/Kind'}
        tag:
          type: string
          nullable: true
          x-validations: ["Size <=50", "Immutable"]
        secret:
          type: string
          x-sensitive: true
    Items:
      type: array
      items: {$ref: '#/components/schemas/Item'}
    Kind:
      type: string
      enum: [active, archived, deleted]
    Event:
      oneOf:
        - {$ref: '#/components/schemas/CreatedEvent'}
        - {$ref: '#/components/schemas/DeletedEvent'}
    CreatedEvent:
      type: object
      required: [type, itemId]
      properties:
        type: {type: string, enum: [created]}
        itemId:
          type: string
          x-validations: ["Size >=1"]
    DeletedEvent:
      type: object
      required: [type, itemId]
      properties:
        type: {type: string, enum: [deleted]}
        itemId:
          type: string
          x-validations: ["Size >=1"]